requirements.txt
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
diff-match-patch==20230430
distlib==0.3.7
Django==4.2.4
django-colorfield==0.11.0
django-import-export==4.0.3
django-js-asset==2.1.0
django-mptt==0.14.0
filelock==3.12.3
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
Pillow==10.0.0
platformdirs==3.10.0
# psycopg2==2.9.7
python-telegram-bot==20.8
requests==2.31.0
sniffio==1.3.1
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
urllib3==2.2.1
virtualenv==20.24.4
whitenoise==6.7.0
webcolors==24.8.0
